// frontend/src/ipc/generated/LifeLinkTargetView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LifeLinkDocumentKind } from "./LifeLinkDocumentKind";

export type LifeLinkTargetView = { node_id: string, title: string, short_description: string, icon_key: string, document_kind: LifeLinkDocumentKind, breadcrumb: string, };